+++ src/Knet.Kudu.Client/Internal/AvlTreeExtensions.cs
@@ -5,19 +5,8 @@ namespace Knet.Kudu.Client.Internal
 {
     public static class AvlTreeExtensions
     {
-        public static RemoteTablet GetFloor(
-            this AvlTree avlTree, ReadOnlySpan<byte> partitionKey)
-        {
-            avlTree.SearchLeftRight(
-                partitionKey,
-                out RemoteTablet left,
-                out _);
-
-            return left;
-        }
-
         public static TabletLocationEntry FloorEntry(
-            this AvlTree2 avlTree, ReadOnlySpan<byte> partitionKey)
+            this AvlTree avlTree, ReadOnlySpan<byte> partitionKey)
         {
             avlTree.SearchLeftRight(
                     partitionKey,
